+++ apps/web/components/layout/navbar.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { SimplistIcon } from "@/components/simplist-icon";
+import { SimplistIcon, SimplistIconThemed } from "@/components/simplist-icon";
 import { buttonVariants } from "@simplist/ui/components/button";
 import { Bars, Xmark } from "@gravity-ui/icons";
 import Link from "next/link";
@@ -23,6 +23,11 @@ const navLinks = [
     label: "Documentation",
     external: true,
   },
+  // {
+  //   href: "/posts",
+  //   label: "Posts",
+  //   external: false
+  // }
 ];
 
 export const Navbar: FC<NavbarProps> = ({ sticky = true }) => {
@@ -36,12 +41,12 @@ export const Navbar: FC<NavbarProps> = ({ sticky = true }) => {
         className={cn(
           "mx-auto max-w-4xl",
           "rounded-2xl border border-white/10",
-          "bg-background/35 backdrop-blur-xl",
+          "dark:bg-background/35 bg-black/5 backdrop-blur-xl",
         )}
       >
         <div className="flex h-14 items-center justify-between px-5">
           {/* Logo - gauche */}
-          <SimplistIcon />
+          <SimplistIconThemed />
 
           {/* Liens + CTA - droite */}
           <div className="flex items-center gap-6">