+++ apps/web/components/simplist-icon.tsx
@@ -1,3 +1,4 @@
+import { IconThemed } from "@simplist/ui/components/icon-themed";
 import Image from "next/image";
 import Link from "next/link";
 
@@ -14,4 +15,8 @@ export const SimplistIcon = () => {
       />
     </Link>
   );
+};
+
+export const SimplistIconThemed = () => {
+  return <IconThemed light={<SimplistIcon />} dark={<SimplistIcon />} />;
 };